examples/run_for_every/pyproject.toml
[project]
name = "run_for_every"
version = "0.1.0"
description = "Example Junjo application demonstrating RunForEvery"
requires-python = ">=3.11"
dependencies = [
  "google-genai>=1.3.0",
  "junjo",                                             # TODO: Convert to github link when package is published
  "load-dotenv>=0.1.0",
  "openinference-instrumentation-google-genai>=0.1.1",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff~=0.9.5"]

[tool.ruff]
# https://docs.astral.sh/ruff/settings/
line-length = 120        # Default, compatible with Black
target-version = "py311" # Or your project's Python version
indent-width = 4         # Default, compatible with Black

# Formatting options (optional - Ruff's formatter is Black-compatible by default)
[tool.ruff.format]
quote-style = "double" # Or "single"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "C", "B"]
ignore = ["B006"]

[tool.ruff.lint.per-file-ignores]
"tests/junjo/test_*.py" = ["D", "F401"] # ignore docstring rules in tests

[tool.uv.sources]
junjo = { workspace = true }

[dependency-groups]
dev = ["ruff>=0.9.5"]
